python/VTuber_Project/requirements.txt
websocket-client
requests
google-cloud-aiplatform